+++ src/main/java/com/study/petory/domain/user/controller/AuthController.java
@@ -15,7 +15,7 @@ import com.study.petory.common.exception.enums.SuccessCode;
 import com.study.petory.common.response.CommonResponse;
 import com.study.petory.domain.user.dto.TokenResponseDto;
 import com.study.petory.domain.user.entity.Role;
-import com.study.petory.domain.user.service.AuthService;
+import com.study.petory.domain.user.service.AuthServiceImpl;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,7 +24,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class AuthController {
 
-	private final AuthService authService;
+	private final AuthServiceImpl authServiceImpl;
 
 	/**
 	 * [토큰 재발급]
@@ -42,7 +42,7 @@ public class AuthController {
 		@RequestHeader("Authorization-Refresh") String refreshToken
 	) {
 
-		TokenResponseDto tokenResponseDto = authService.reissue(accessToken, refreshToken);
+		TokenResponseDto tokenResponseDto = authServiceImpl.reissue(accessToken, refreshToken);
 
 		return CommonResponse.of(SuccessCode.TOKEN_REISSUE, tokenResponseDto);
 	}
@@ -59,7 +59,7 @@ public class AuthController {
 	@PreAuthorize("hasAnyRole('USER', 'ADMIN')")
 	public ResponseEntity<CommonResponse<Object>> logout(@RequestHeader("Authorization") String bearerToken) {
 
-		authService.logout(bearerToken);
+		authServiceImpl.logout(bearerToken);
 
 		return CommonResponse.of(SuccessCode.USER_LOGOUT);
 	}
@@ -79,7 +79,7 @@ public class AuthController {
 		@RequestParam("role") Role role
 	) {
 
-		List<Role> updatedRoles = authService.addRoleToUser(targetUserId, role);
+		List<Role> updatedRoles = authServiceImpl.addRoleToUser(targetUserId, role);
 
 		return CommonResponse.of(SuccessCode.UPDATED, updatedRoles);
 	}
@@ -99,7 +99,7 @@ public class AuthController {
 		@RequestParam("role") Role role
 	) {
 
-		List<Role> updatedRoles = authService.removeRoleFromUser(targetUserId, role);
+		List<Role> updatedRoles = authServiceImpl.removeRoleFromUser(targetUserId, role);
 
 		return CommonResponse.of(SuccessCode.DELETED, updatedRoles);
 	}
@@ -116,7 +116,7 @@ public class AuthController {
 	@DeleteMapping("/deactivate")
 	public ResponseEntity<CommonResponse<Object>> deactivateUser(@RequestParam Long userId) {
 
-		authService.deactivateUser(userId);
+		authServiceImpl.deactivateUser(userId);
 
 		return CommonResponse.of(SuccessCode.DELETED);
 	}
@@ -134,7 +134,7 @@ public class AuthController {
 	@PostMapping("/restore")
 	public ResponseEntity<CommonResponse<Object>> restoreUser(@RequestParam Long userId) {
 
-		authService.restoreUser(userId);
+		authServiceImpl.restoreUser(userId);
 
 		return CommonResponse.of(SuccessCode.RESTORED);
 	}
